requirements.txt
streamlit
spacy==3.7.2
scikit-learn
pdfminer.six